Fix tests for linux.

+++ Hexa.NET.DirectXTex.Tests/BlobMethods.cs
@@ -14,7 +14,7 @@
         {
             Blob blob = DirectXTex.CreateBlob();
             blob.Initialize(256);
-            Assert.That(blob.GetBufferSize(), Is.EqualTo((ulong)256));
+            Assert.That(blob.GetBufferSize(), Is.EqualTo((nuint)256));
 
             blob.Release();
         }
